@@ -8,48 +8,49 @@
     "test": "vitest"
   },
   "devDependencies": {
-    "@iconify-json/mdi": "^1.1.68",
+    "@iconify-json/mdi": "^1.2.3",
     "@originjs/vite-plugin-commonjs": "^1.0.3",
-    "@radix-ui/react-hover-card": "^1.1.1",
-    "@radix-ui/react-popover": "^1.1.1",
-    "@radix-ui/react-tooltip": "^1.1.2",
+    "@radix-ui/react-hover-card": "^1.1.15",
+    "@radix-ui/react-popover": "^1.1.15",
+    "@radix-ui/react-tooltip": "^1.2.8",
     "@rschedule/core": "^1.5.0",
     "@rschedule/luxon-date-adapter": "^1.5.0",
     "@svgr/core": "^8.1.0",
-    "@types/lodash": "^4.17.7",
-    "@types/luxon": "^3.4.2",
-    "@types/node": "^22.1.0",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^8.0.1",
-    "@typescript-eslint/parser": "^8.0.1",
-    "@vitejs/plugin-react": "^4.3.1",
-    "autoprefixer": "^10.4.20",
-    "c8": "^10.1.2",
+    "@tailwindcss/postcss": "^4.1.13",
+    "@types/lodash": "^4.17.20",
+    "@types/luxon": "^3.7.1",
+    "@types/node": "^24.3.1",
+    "@types/react": "^19.1.12",
+    "@types/react-dom": "^19.1.9",
+    "@typescript-eslint/eslint-plugin": "^8.42.0",
+    "@typescript-eslint/parser": "^8.42.0",
+    "@vitejs/plugin-react": "^5.0.2",
+    "autoprefixer": "^10.4.21",
+    "c8": "^10.1.3",
     "classnames": "^2.5.1",
-    "eslint": "^9.8.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "^2.29.1",
+    "eslint": "^9.35.0",
+    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-plugin-import": "^2.32.0",
     "lodash": "^4.17.21",
-    "luxon": "^3.5.0",
-    "mobx": "^6.13.1",
-    "mobx-react": "^9.1.1",
-    "mobx-state-tree": "^6.0.1",
-    "mobx-utils": "^6.0.8",
-    "nanoid": "^5.0.7",
-    "peggy": "^4.0.3",
+    "luxon": "^3.7.2",
+    "mobx": "^6.13.7",
+    "mobx-react": "^9.2.0",
+    "mobx-state-tree": "^7.0.2",
+    "mobx-utils": "^6.1.1",
+    "nanoid": "^5.1.5",
+    "peggy": "^5.0.6",
     "pegjs-backtrace": "^0.2.1",
-    "postcss": "^8.4.41",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-spring": "^9.7.4",
+    "postcss": "^8.5.6",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-spring": "^10.0.1",
     "rrule": "^2.8.1",
-    "tailwindcss": "^3.4.9",
-    "typescript": "^5.5.4",
-    "unplugin-auto-import": "^0.18.2",
-    "unplugin-icons": "^0.19.2",
-    "vite": "^5.4.0",
-    "vitest": "^2.0.5"
+    "tailwindcss": "^4.1.13",
+    "typescript": "^5.9.2",
+    "unplugin-auto-import": "^20.1.0",
+    "unplugin-icons": "^22.2.0",
+    "vite": "^7.1.4",
+    "vitest": "^3.2.4"
   },
   "browserslist": {
     "production": [
@@ -64,6 +65,7 @@
     ]
   },
   "dependencies": {
-    "emoji-from-word": "^1.2.12"
+    "@radix-ui/themes": "^3.2.1",
+    "emoji-from-word": "^1.2.13"
   }
 }
