@@ -1,11 +1,17 @@
-// Generated by 'unplugin-auto-import'
-// We suggest you to commit this file into source control
+/* eslint-disable */
+/* prettier-ignore */
+// @ts-nocheck
+// noinspection JSUnusedGlobalSymbols
+// Generated by unplugin-auto-import
+// biome-ignore lint: disable
+export {}
 declare global {
   const IconMdiCalendarBlank: typeof import('~icons/mdi/calendar-blank.jsx')['default']
   const IconMdiClockOutline: typeof import('~icons/mdi/clock-outline.jsx')['default']
   const IconMdiCog: typeof import('~icons/mdi/cog.jsx')['default']
   const IconMdiContentCopy: typeof import('~icons/mdi/content-copy.jsx')['default']
   const IconMdiContentPaste: typeof import('~icons/mdi/content-paste.jsx')['default']
+  const IconMdiPlaylistPlus: typeof import('~icons/mdi/playlist-plus.jsx')['default']
   const IconMdiTimerOutline: typeof import('~icons/mdi/timer-outline.jsx')['default']
   const IconMdiTrashCan: typeof import('~icons/mdi/trash-can.jsx')['default']
   const IconMdiTrashCanOutline: typeof import('~icons/mdi/trash-can-outline.jsx')['default']
@@ -13,4 +19,3 @@ declare global {
   const IconMdiWeatherNight: typeof import('~icons/mdi/weather-night.jsx')['default']
   const IconMdiWeatherSunny: typeof import('~icons/mdi/weather-sunny.jsx')['default']
 }
-export {}
