@@ -14,22 +14,20 @@ import Input from "./Input";
 
 import Agenda from "./Agenda";
 
-// export const Colors = ["#06b6d4", "#0ea5e9", "#10b981", "#14b8a6", "#22c55e", "#3b82f6", "#6366f1", "#84cc16", "#8b5cf6", "#a855f7", "#d946ef", "#eab308", "#ec4899", "#ef4444", "#f43f5e", "#f97316"]
-// export const Colors = ["#ffbe0b","#fd8a09","#fb5607","#fd2b3b","#ff006e","#c11cad","#8338ec","#5f5ff6","#3a86ff"]
-// export const Colors = ["#ff009a","#ff0000","#ff9100","#ffce32","#00c800","#00cccc","#3b78d8","#9900ff","#ec00ec"]
-// export const Colors = ["#e63989","#653d8b","#f4e508","#f3aa05","#81d0dc","#e4191f","#b0151d","#ea6c39","#5fb264"]
-// export const Colors = ["#54478c", "#2c699a", "#048ba8", "#0db39e", "#16db93", "#83e377", "#b9e769", "#efea5a", "#f1c453", "#f29e4c"]
-export const Colors = [
-  "#33a8c7",
-  "#52e3e1",
-  "#a0e426",
-  "#fdf148",
-  "#ffab00",
-  "#f77976",
-  "#f050ae",
-  "#d883ff",
-  "#9336fd",
-];
+const colorPalettes = {
+  base16: [
+    "var(--base08)",
+    "var(--base09)",
+    "var(--base0A)",
+    "var(--base0B)",
+    "var(--base0C)",
+    "var(--base0D)",
+    "var(--base0E)",
+    "var(--base0F)",
+  ],
+};
+
+export const Colors = colorPalettes.base16;
 
 interface StoreVolatileProps {
   hoveredTask: ITask | null;
@@ -54,9 +52,7 @@ const TimeOfTheDay = t
   })
   .actions((self) => ({
     set(name: string, hour: number) {
-      if (name in self) {
-        self[name as keyof ITimeOfTheDaySnapshotIn] = hour;
-      }
+      (self as any)[name] = hour;
     },
   }));
 
@@ -145,9 +141,9 @@ const Store = t
     },
 
     get contexts() {
-      return _.uniq(self.tasks.flatMap((task) => task.contexts)).filter(
-        Boolean
-      );
+      return _.uniq(self.tasks.flatMap((task) => task.contexts))
+        .filter(Boolean)
+        .sort();
     },
 
     get asList() {
@@ -165,11 +161,16 @@ const Store = t
 
         occurrences.forEach((occurrence) => {
           const day = occurrence.startOf("day").toISODate();
-          const existing = result.get(day) ?? [];
-          result.set(
-            day,
-            _.uniqBy([...existing, new Occurrence(occurrence, task)], "task.id")
-          );
+          if (day) {
+            const existing = result.get(day) ?? [];
+            result.set(
+              day,
+              _.uniqBy(
+                [...existing, new Occurrence(occurrence, task)],
+                "task.id"
+              )
+            );
+          }
         });
       });
 
@@ -177,16 +178,17 @@ const Store = t
     },
 
     getOccurrencesAtDay(day: DateTime): Occurrence[] {
-      return _.sortBy(
-        this.occurrencesByDay.get(day.startOf("day").toISODate()) ?? [],
-        "date"
-      );
+      const dayKey = day.startOf("day").toISODate();
+      if (!dayKey) return [];
+      return _.sortBy(this.occurrencesByDay.get(dayKey) ?? [], "date");
     },
 
-    getContextColor(context?: string) {
-      if (context)
-        return Colors[this.contexts.indexOf(context) % Colors.length];
-      return self.useDarkMode ? "#444" : "#ccc";
+    getContextColor(context?: string): string {
+      if (!context) {
+        return "var(--base03)";
+      }
+      const contextIndex = this.contexts.indexOf(context);
+      return Colors[contextIndex % Colors.length];
     },
   }))
   .actions((self) => ({
