@@ -2,7 +2,26 @@ module.exports = {
   content: ["src/**/*.{js,jsx,ts,tsx}"],
   darkMode: "class",
   theme: {
-    extend: {},
+    extend: {
+      colors: {
+        'base-00': 'var(--base00)',
+        'base-01': 'var(--base01)',
+        'base-02': 'var(--base02)',
+        'base-03': 'var(--base03)',
+        'base-04': 'var(--base04)',
+        'base-05': 'var(--base05)',
+        'base-06': 'var(--base06)',
+        'base-07': 'var(--base07)',
+        'base-08': 'var(--base08)',
+        'base-09': 'var(--base09)',
+        'base-0A': 'var(--base0A)',
+        'base-0B': 'var(--base0B)',
+        'base-0C': 'var(--base0C)',
+        'base-0D': 'var(--base0D)',
+        'base-0E': 'var(--base0E)',
+        'base-0F': 'var(--base0F)',
+      },
+    },
   },
   plugins: [],
 }
